+++ example/lib/l10n/localization.dart
@@ -3,7 +3,6 @@
 
 abstract class LocaleKeys {
   static const hiMessage = 'hiMessage';
-  static const increment = 'increment';
   static const itemCounter = 'itemCounter';
   static const languageFlag = 'languageFlag';
 }
@@ -13,14 +12,12 @@ final initialTranslations = [
     "@@locale": "en",
     LocaleKeys.languageFlag: "🇺🇸",
     LocaleKeys.hiMessage: "Hi {name}!",
-    LocaleKeys.increment: "Increment",
     LocaleKeys.itemCounter: "Currently there { count, plural, =0{are no items} =1{is one item} other{are # items}} in this app",
   },
   {
     "@@locale": "es",
     LocaleKeys.languageFlag: "🇪🇸",
     LocaleKeys.hiMessage: "¡Hola {name}!",
-    LocaleKeys.increment: "Incremento",
     LocaleKeys.itemCounter: "Actualmente hay { count, plural, =0{no hay elementos} =1{un elemento} other{# elementos}} en esta aplicación",
   },
 ];